+++ src/api-c/OpenKit-c.cxx
@@ -703,6 +703,11 @@ extern "C" {
 		return handle;
 	}
 
+	SessionHandle* createSessionWithAutoIpDetermination(struct OpenKitHandle* openKitHandle)
+	{
+		return createSession(openKitHandle, nullptr);
+	}
+
 	void endSession(SessionHandle* sessionHandle)
 	{
 		// Sanity